+++ test/fixtures/conditional-arrow.js
@@ -1,0 +1,7 @@
+var StateMachine = require('fsm-as-promised');
+
+StateMachine({
+  events: [
+    { name: 'multiple', from: 'none', to: ['one', 'two'], condition: () => { return 0 } }
+  ]
+});